backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.35
aiosqlite==0.20.0
alembic==1.13.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
httpx==0.27.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
loguru==0.7.2
email-validator==2.1.1
numpy==2.1.3
pandas==2.2.3
scikit-learn==1.5.2
torch==2.5.1
pytest==8.3.3

